refactor: split LFM2-VL loader metadata helpers
Move the processor_config parsing, tiling policy validation, marker-token resolution, and loader unit tests out of the LFM2-VL loader so the loader remains focused on model assembly and stays below the file-size limit.

Behavior is unchanged: the new metadata module exposes the same helpers to the loader, and the tests are the previous loader regressions moved into a sibling _tests file.
